+++ src/main/java/net/minecraft/world/chunk/Chunk.java
@@ -1150,7 +1150,7 @@ public class Chunk
 
     public Random getRandomWithSeed(long par1)
     {
-        return new Random(this.worldObj.getSeed() + (long) ((long) this.xPosition * this.xPosition * 4987142) + (long) (this.xPosition * 5947611L) + (long) ((long) this.zPosition * this.zPosition) * 4392871L + (long) (this.zPosition * 389711L) ^ par1);
+        return new Random(this.worldObj.getSeed() + ((long) this.xPosition * this.xPosition * 4987142) + (this.xPosition * 5947611L) + (long) this.zPosition * this.zPosition * 4392871L + (this.zPosition * 389711L) ^ par1);
     }
 
     public boolean isEmpty()
